pyproject.toml
[project]
name = "pokemon-card-detector"
version = "1.0.0"
description = "AI-powered Pokemon card authenticity detector"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=1.24.3",
    "opencv-python>=4.8.1.78",
    "pillow>=10.0.1",
    "psycopg2-binary>=2.9.7",
    "pytesseract>=0.3.10",
    "qrcode>=8.2",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.23",
    "werkzeug>=3.1.3",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pyjwt[crypto]>=2.10.1",
    "cryptography>=45.0.5",
    "jwcrypto>=1.5.6",
    "tensorflow>=2.14.0",
]
